Promote the CLI out of utilities
