+++ examples/convert_from_pt_to_pl/image_classifier_5_lightning_datamodule.py
@@ -25,7 +25,7 @@ from torchmetrics import Accuracy
 from pytorch_lightning import cli_lightning_logo, LightningDataModule, LightningModule
 from pytorch_lightning.demos.boring_classes import Net
 from pytorch_lightning.demos.mnist_datamodule import MNIST
-from pytorch_lightning.utilities.cli import LightningCLI
+from pytorch_lightning.cli import LightningCLI
 
 DATASETS_PATH = path.join(path.dirname(__file__), "..", "..", "Datasets")
