+++ examples/pl_domain_templates/computer_vision_fine_tuning.py
@@ -56,7 +56,7 @@ from torchvision.datasets.utils import download_and_extract_archive
 
 from pytorch_lightning import cli_lightning_logo, LightningDataModule, LightningModule
 from pytorch_lightning.callbacks.finetuning import BaseFinetuning
-from pytorch_lightning.utilities.cli import LightningCLI
+from pytorch_lightning.cli import LightningCLI
 from pytorch_lightning.utilities.rank_zero import rank_zero_info
 
 log = logging.getLogger(__name__)
